requirements.txt
aiomongo
discord.py>=2.0.0
python-dotenv
aiofiles